[assistant/configuration] - feature: implement YAML export functionality for agent configurations
 - Created a new service to export agent configurations to a YAML file format
 - Refactored existing code to centralize YAML export logic into a new dedicated module
 - Updated the API endpoint to use the new YAML export service for downloading agent configurations
 - Removed duplicated and now redundant code from the previous implementation of YAML export in API endpoint
